feat: migrate build to docusaurus with the open-rpc plugin.
This change migrates the build from gatsby to docusaurus.
The build was growing rusty and was blocking serious devex
improvements.

This change enables, documentation search, fork/eip/category tagging, support for error groups, copying as markdown, and standardized docusaurus support.

The build also writes documentation do a documentation folder as
a commit this changes the previous method, while it does add a small amount of incremental noise, it improves the devex, as you can pinpoint rendering/typo issues etc.. more directly and link to them.

One thing to note is that the documentation is auto generated and
should not be edited directly.